refactor(cli): Share config/list/rules/stats commands across platforms
Restore the platform-agnostic CLI entrypoints and keep only the Stats struct Linux/Windows specific. Add Linux DialPipe and hostname helpers so the shared commands compile.
