+++ cmd/fibratus/app/rules/list.go
@@ -1,5 +1,3 @@
-//go:build windows
-
 /*
  * Copyright 2021-2022 by Nedim Sabic Sabic
  * https://www.fibratus.io
