+++ internal/bootstrap/bootstrap_windows.go
